fix(frontend): restore three lint rules lost in the oxlint port, deny warnings
The oxlint config only enabled the `correctness` category, which silently
dropped three rules the old ESLint config had. Two were real regressions:

- `react-hooks/rules-of-hooks` -> `react/hooks` lives in `suspicious`, so it
  was off. It immediately caught a genuine hazard the anti-slop test-seam work
  introduced: `FleetQueuePanel` was injecting *hooks* through its optional
  `deps` prop, so hook identity was not statically provable. The panel now
  calls `useFleetQueue`/`useFleetSummary` directly and its test seeds a real
  `QueryClient` instead — the pattern `printers-list.test.tsx` already uses.
- `no-empty` (which the old config had as an error with `allowEmptyCatch`) and
  `react-refresh/only-export-components` were both off.

`only-export-components` is fixed by splitting the modules rather than by
allowlisting the export names, so Fast Refresh genuinely works:

- `lib/auth-context.tsx` -> `auth-context.ts` (context/hook/types, 17
  importers untouched) + `auth-provider.tsx`
- `lib/navigation.tsx` -> `navigation.ts` (hooks) + `link.tsx` (component)
- `lib/mobile-filter-context.tsx` -> `.ts` (context/hook) + provider
- `MODEL_DND_MIME` -> `lib/model-dnd.ts`
- `inputClasses` -> `packages/ui/src/lib/input-classes.ts`, re-exported from
  the package index so `@printstash/ui`'s public surface is unchanged
- the g-code parser -> `lib/gcode.ts` (it had no consumer outside the viewer)

`pnpm lint` is now `oxlint --deny-warnings`, so a warning fails the build
locally and in CI rather than scrolling past. Verified by temporarily setting a
rule to `warn` and confirming a non-zero exit.

Also: `oxfmt` formats `.oxlintrc.json`, so the checked-in config had to be
oxfmt-clean or `pnpm format:check` would have failed CI. And the CI job now
installs Playwright browsers *after* typecheck/lint/format/unit/build, so a
lint failure fails in seconds instead of after a browser download.
