+++ frontend/packages/ui/src/components/input.tsx
@@ -1,9 +1,7 @@
 import * as React from "react";
 
+import { inputClasses } from "../lib/input-classes";
 import { cn } from "../lib/utils";
-
-export const inputClasses =
-  "flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 aria-[invalid=true]:border-destructive aria-[invalid=true]:focus-visible:ring-destructive";
 
 export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
