+++ frontend/src/components/sidebar-nav.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { Link } from "@/lib/navigation";
+import { Link } from "@/lib/link";
 import { usePathname } from "@/lib/navigation";
 import {
   BookOpen,
