+++ frontend/src/lib/model-dnd.ts
@@ -1,0 +1,3 @@
+// MIME type carrying a dragged model's id, so collection drop targets can tell
+// an internal "move model" drag apart from an OS file-upload drag.
+export const MODEL_DND_MIME = "application/x-printstash-model";
