+++ frontend/src/root-layout.tsx
@@ -2,7 +2,7 @@ import { Outlet } from "react-router-dom";
 
 import { AppShell } from "@/components/app-shell";
 import { SetupGate } from "@/components/setup-gate";
-import { AuthProvider } from "@/lib/auth-context";
+import { AuthProvider } from "@/lib/auth-provider";
 
 /**
  * Root layout: mirrors the old Next `app/layout.tsx` provider stack, with
